Change storys to .txt Files.
