+++ src/assets/Little_Red_HenText.txt
@@ -1,0 +1,1 @@
+Once upon a time there was a little red hen who lived on a farm. /n Early one morning she woke up and went outside. There she found some corn. /n “Who will help me plant the corn?” said the little red hen. /n “Not I,” said the bull. /n “Not I,” said the cat. /n “Not I,” said the rat. /n “Oh very well, I’ll do it myself,” said the little red hen – and so she did! /n “Who will help me water the corn?” said the little red hen. /n “Not I,” said the bull. /n “Not I,” said the cat. /n “Not I,” said the rat. /n “Oh very well, I’ll do it myself,” said the little red hen – and so she did! /n “Who will help me cut the corn?” said the little red hen. /n “Not I,” said the bull. /n “Not I,” said the cat. /n “Not I,” said the rat. /n “Oh very well, I’ll do it myself,” said the little red hen – and so she did! /n “Who will help me carry the corn to the mill?” said the little red hen. /n “Not I,” said the bull. /n “Not I,” said the cat. /n “Not I,” said the rat. /n “Oh very well, I’ll do it myself,” said the little red hen – and so she did! /n “Who will help me grind the corn?” said the little red hen. /n “Not I,” said the bull. /n “Not I,” said the cat. /n “Not I,” said the rat. /n “Oh very well, I’ll do it myself,” said the little red hen – and so she did! /n “Who will help me knead the bread?” said the little red hen. /n “Not I,” said the bull. /n “Not I,” said the cat. /n “Not I,” said the rat. /n “Oh very well, I’ll do it myself,” said the little red hen – and so she did! /n “Who will help me bake the bread?” said the little red hen. /n “Not I,” said the bull. /n “Not I,” said the cat. /n “Not I,” said the rat. /n “Oh very well, I’ll do it myself,” said the little red hen – and so she did! /n “Who will help me eat the bread?” said the little red hen. /n “I will,” said the bull. /n “I will,” said the cat. /n “I will,” said the rat. /n “Oh no you won’t. I’ll eat it myself,” said the little red hen – and so she did!
