+++ src/assets/PeterRabbitText.txt
@@ -1,0 +1,1 @@
+Once upon a time there were four little rabbits, and their names were Flopsy, Mopsy, Cotton-tail and Peter. They lived with their mother in a sand-bank, underneath the root of a very big fir tree. /n "Now, my dears," said old Mrs. Rabbit one morning, "You may go into the fields or down the lane, but don't go into Mr. McGregor's garden. Your father had an accident there; he was put in a pie by Mrs. McGregor." /n "Now run along and don't get into mischief. I am going out." Then old Mrs. Rabbit took a basket and her umbrella and went through the wood to the baker's. She bought a loaf of brown bread and five currant buns. /n Flopsy, Mopsy and Cotton-tail who were good little bunnies went down the lane together. To gather blackberries. But Peter who was very naughty, ran straight away to Mr. McGregor's garden and squeezed under the gate! /n First he ate some lettuces and some French beans. /n And then he ate some radishes and then, feeling rather sick, he went to look for some parsley. /n But round the end of a cucumber frame, whom should he meet but Mr. McGregor! /n Mr. McGregor was on his hands and knees planting out young cabbages, but he jumped up and ran after Peter, waving a rake and calling out "Stop thief!" /n Peter was most dreadfully frightened; he rushed all over the garden, for he had forgotten the way back to the gate. He lost one shoe among the cabbages, and the other amongst the potatoes. After losing them, he ran on four legs and went faster. So that I think he might have got away altogether if he had not unfortunately run into a gooseberry net. And got caught by the large buttons on his jacket. /n It was a blue jacket with brass buttons, quite new. Peter gave himself up for lost and shed big tears; But his sobs were overheard by some friendly sparrows, who flew to him in great excitement and implored him to exert himself. /n Mr. McGregor came up with a sieve which he intended to pop on the top of Peter, but Peter wriggled out just in time. Leaving his jacket behind him. /n He rushed into the tool-shed and-- Jumped into a can. It would have been a beautiful thing to hide in, if it had not had so much water in it. /n Mr. McGregor was quite sure that Peter was somewhere in the tool-shed, perhaps hidden underneath a flower-pot. He began to turn them over carefully, looking under each. /n Presently Peter sneezed "Kertyschoo!" /n Mr. McGregor was after him in no time, and tried to put his foot upon Peter, who Jumped out of a window, upsetting three plants. /n Peter sat down to rest; he was out of breath and trembling with fright, and he had not the least idea which way to go. Also he was very damp with sitting in that can. /n After a time he began to wander about, going lippity-- lippity-- not very fast and looking all around. He found a door in a wall; but it was locked and there was no room for a fat little rabbit to squeeze underneath. /n An old mouse was running in and out over the stone doorstep, carrying peas and beans to her family in the wood. Peter asked her the way to the gate but she had such a large pea in her mouth she could not answer. She only shook her head at him. /n Peter began to cry. /n Then he tried to find his way straight across the garden, but he became more and more puzzled. Presently he came to a pond where Mr. McGregor filled his water-cans. /n A white cat was staring at some gold-fish; she sat very, very still, but now and then the tip of her tail twitched as if it were alive. /n Peter thought it best to go away without speaking to her. He had heard about cats from his cousin, little Benjamin Bunny. /n He went back towards the tool-shed, but suddenly, quite close to him, he heard the noise of a hoe-scritch, scratch, scratch, scritch. /n Peter scuttered underneath the bushes, but presently as nothing happened, he came out and climbed upon a wheelbarrow, and peeped over. /n The first thing he saw was Mr. McGregor hoeing onions. His back was turned towards Peter and beyond him was the gate! /n Peter got down very quietly off the wheel-barrow and started running as fast as he could go, along a straight walk behind some black currant bushes. /n Mr. McGregor caught sight of him at the corner, but Peter did not care. He slipped underneath the gate and was safe at last in the wood outside the garden. /n Mr. McGregor hung up the little jacket and the shoes for a scare-crow to frighten the blackbirds. /n Peter never stopped running or looked behind him till he got home to the big fir-tree. He was so tired that he flopped down upon the nice soft sand on the floor of the rabbit hole, and shut his eyes. /n His mother was busy cooking; she wondered what he had done with his clothes. It was the second little jacket and pair of shoes that Peter had lost in a fortnight! /n I am sorry to say that Peter was not very well during the evening. His mother put him to bed and made some camomile tea; and she gave a dose of it to Peter! /n "One teaspoonful to be taken at bedtime." But -- Flopsy, Mopsy and Cottontail had bread and milk and blackberries for supper.
