--- src/assets/StoryText.js
@@ -1,64 +1,0 @@
-/**To add a new book to the library:
-    1. Import the cover art that you would like to appear on the bookshelf
-    2. Put book text in a variable
-    3. Create an object following the 'Text, SearchTerms, Cover' format
-        -Every Story must have a search term of either 'early readers' or 'advanced readers' based on reading level
-    4. Put book in export object
-    5. The end
-*/
-
-import peterRabbitCover from '../assets/peterRabbitCover.jpg';
-import tortHareCover from '../assets/TortoiseHareCover.jpg';
-import littleRedHenCover from '../assets/littleRedHenCover.jpg';
-import theBlackCatCover from '../assets/theblackcatCover.jpg';
-import RedDeathCover from '../assets/RedDeathCover.jpg';
-
-let PeterRabbitText = `Once upon a time there were four little rabbits, and their names were Flopsy, Mopsy, Cotton-tail and Peter. They lived with their mother in a sand-bank, underneath the root of a very big fir tree. /n "Now, my dears," said old Mrs. Rabbit one morning, "You may go into the fields or down the lane, but don't go into Mr. McGregor's garden. Your father had an accident there; he was put in a pie by Mrs. McGregor." /n "Now run along and don't get into mischief. I am going out." Then old Mrs. Rabbit took a basket and her umbrella and went through the wood to the baker's. She bought a loaf of brown bread and five currant buns. /n Flopsy, Mopsy and Cotton-tail who were good little bunnies went down the lane together. To gather blackberries. But Peter who was very naughty, ran straight away to Mr. McGregor's garden and squeezed under the gate! /n First he ate some lettuces and some French beans. /n And then he ate some radishes and then, feeling rather sick, he went to look for some parsley. /n But round the end of a cucumber frame, whom should he meet but Mr. McGregor! /n Mr. McGregor was on his hands and knees planting out young cabbages, but he jumped up and ran after Peter, waving a rake and calling out "Stop thief!" /n Peter was most dreadfully frightened; he rushed all over the garden, for he had forgotten the way back to the gate. He lost one shoe among the cabbages, and the other amongst the potatoes. After losing them, he ran on four legs and went faster. So that I think he might have got away altogether if he had not unfortunately run into a gooseberry net. And got caught by the large buttons on his jacket. /n It was a blue jacket with brass buttons, quite new. Peter gave himself up for lost and shed big tears; But his sobs were overheard by some friendly sparrows, who flew to him in great excitement and implored him to exert himself. /n Mr. McGregor came up with a sieve which he intended to pop on the top of Peter, but Peter wriggled out just in time. Leaving his jacket behind him. /n He rushed into the tool-shed and-- Jumped into a can. It would have been a beautiful thing to hide in, if it had not had so much water in it. /n Mr. McGregor was quite sure that Peter was somewhere in the tool-shed, perhaps hidden underneath a flower-pot. He began to turn them over carefully, looking under each. /n Presently Peter sneezed "Kertyschoo!" /n Mr. McGregor was after him in no time, and tried to put his foot upon Peter, who Jumped out of a window, upsetting three plants. /n Peter sat down to rest; he was out of breath and trembling with fright, and he had not the least idea which way to go. Also he was very damp with sitting in that can. /n After a time he began to wander about, going lippity-- lippity-- not very fast and looking all around. He found a door in a wall; but it was locked and there was no room for a fat little rabbit to squeeze underneath. /n An old mouse was running in and out over the stone doorstep, carrying peas and beans to her family in the wood. Peter asked her the way to the gate but she had such a large pea in her mouth she could not answer. She only shook her head at him. /n Peter began to cry. /n Then he tried to find his way straight across the garden, but he became more and more puzzled. Presently he came to a pond where Mr. McGregor filled his water-cans. /n A white cat was staring at some gold-fish; she sat very, very still, but now and then the tip of her tail twitched as if it were alive. /n Peter thought it best to go away without speaking to her. He had heard about cats from his cousin, little Benjamin Bunny. /n He went back towards the tool-shed, but suddenly, quite close to him, he heard the noise of a hoe-scritch, scratch, scratch, scritch. /n Peter scuttered underneath the bushes, but presently as nothing happened, he came out and climbed upon a wheelbarrow, and peeped over. /n The first thing he saw was Mr. McGregor hoeing onions. His back was turned towards Peter and beyond him was the gate! /n Peter got down very quietly off the wheel-barrow and started running as fast as he could go, along a straight walk behind some black currant bushes. /n Mr. McGregor caught sight of him at the corner, but Peter did not care. He slipped underneath the gate and was safe at last in the wood outside the garden. /n Mr. McGregor hung up the little jacket and the shoes for a scare-crow to frighten the blackbirds. /n Peter never stopped running or looked behind him till he got home to the big fir-tree. He was so tired that he flopped down upon the nice soft sand on the floor of the rabbit hole, and shut his eyes. /n His mother was busy cooking; she wondered what he had done with his clothes. It was the second little jacket and pair of shoes that Peter had lost in a fortnight! /n I am sorry to say that Peter was not very well during the evening. His mother put him to bed and made some camomile tea; and she gave a dose of it to Peter! /n "One teaspoonful to be taken at bedtime." But -- Flopsy, Mopsy and Cottontail had bread and milk and blackberries for supper.`
-let Tortoise_and_the_HareText = `One day a hare was bragging about how fast he could run. He bragged and bragged and even laughed at the tortoise, who was so slow. /n The tortoise stretched out his long neck and challenged the hare to a race, which, of course, made the hare laugh. /n "My, my, what a joke!" thought the hare. "A race, indeed, a race. Oh! what fun! My, my! a race, of course, Mr. Tortoise, we shall race!" said the hare. /n The forest animals met and mapped out the course. The race began, and the hare, being such a swift runner, soon left the tortoise far behind. /n About halfway through the course, it occurred to the hare that he had plenty of time to beat the slow trodden tortoise. "Oh, my!" thought the hare, "I have plenty of time to play in the meadow here." And so he did. /n After the hare finished playing, he decided that he had time to take a little nap. "I have plenty of time to beat that tortoise," he thought. And he cuddle up against a tree and dozed. /n The tortoise, in the meantime, continued to plod on, albeit, it ever so slowly. He never stopped, but took one good step after another. /n The hare finally woke from his nap. "Time to get going," he thought. And off he went faster than he had ever run before! He dashed as quickly as anyone ever could up to the finish line, where he met the tortoise, who was patiently awaiting his arrival. /n Slow and steady wins the race.`
-let Little_Red_HenText = `Once upon a time there was a little red hen who lived on a farm. /n Early one morning she woke up and went outside. There she found some corn. /n “Who will help me plant the corn?” said the little red hen. /n “Not I,” said the bull. /n “Not I,” said the cat. /n “Not I,” said the rat. /n “Oh very well, I’ll do it myself,” said the little red hen – and so she did! /n “Who will help me water the corn?” said the little red hen. /n “Not I,” said the bull. /n “Not I,” said the cat. /n “Not I,” said the rat. /n “Oh very well, I’ll do it myself,” said the little red hen – and so she did! /n “Who will help me cut the corn?” said the little red hen. /n “Not I,” said the bull. /n “Not I,” said the cat. /n “Not I,” said the rat. /n “Oh very well, I’ll do it myself,” said the little red hen – and so she did! /n “Who will help me carry the corn to the mill?” said the little red hen. /n “Not I,” said the bull. /n “Not I,” said the cat. /n “Not I,” said the rat. /n “Oh very well, I’ll do it myself,” said the little red hen – and so she did! /n “Who will help me grind the corn?” said the little red hen. /n “Not I,” said the bull. /n “Not I,” said the cat. /n “Not I,” said the rat. /n “Oh very well, I’ll do it myself,” said the little red hen – and so she did! /n “Who will help me knead the bread?” said the little red hen. /n “Not I,” said the bull. /n “Not I,” said the cat. /n “Not I,” said the rat. /n “Oh very well, I’ll do it myself,” said the little red hen – and so she did! /n “Who will help me bake the bread?” said the little red hen. /n “Not I,” said the bull. /n “Not I,” said the cat. /n “Not I,” said the rat. /n “Oh very well, I’ll do it myself,” said the little red hen – and so she did! /n “Who will help me eat the bread?” said the little red hen. /n “I will,” said the bull. /n “I will,” said the cat. /n “I will,” said the rat. /n “Oh no you won’t. I’ll eat it myself,” said the little red hen – and so she did!`
-let The_Black_CatText = `Tomorrow I die. Tomorrow I die, and today I want to tell the world what happened and thus perhaps free my soul from the horrible weight which lies upon it. /n But listen! Listen, and you shall hear how I have been destroyed. /n When I was a child, I had a natural goodness of soul which led me to love animals — all kinds of animals, but especially those animals we call pets, animals which have learned to live with men and share their homes with them. There is something in the love of these animals which speaks directly to the heart of the man who has learned from experience how uncertain and changeable is the love of other men. /n I was quite young when I married. You will understand the joy I felt to find that my wife shared with me my love for animals. Quickly she got for us several pets of the most likeable kind. We had birds, some goldfish, a fine dog, and a cat. /n The cat was a beautiful animal, of unusually large size, and entirely black. I named the cat Pluto, and it was the pet I liked best. I alone fed it, and it followed me all around the house. It was even with difficulty that I stopped it from following me through the streets. /n Our friendship lasted, in this manner, for several years, during which, however, my own character became greatly changed. I began to drink too much wine and other strong drinks. /n As the days passed I became less loving in my manner; I became quick to anger; I forgot how to smile and laugh. My wife — yes, and my pets, too, all except the cat — were made to feel the change in my character. /n One night I came home quite late from the inn, where I now spent more and more time drinking. Walking with uncertain step, I made my way with effort into the house. As I entered I saw — or thought I saw — that Pluto, the cat, was trying to stay out of my way, to avoid me. This action, by an animal which I had thought still loved me, made me angry beyond reason. My soul seemed to fly from my body. I took a small knife out of my coat and opened it. Then I took the poor animal by the neck and with one quick movement I cut out one of its fear-filled eyes! /n Slowly the cat got well. The hole where its eye had been was not a pretty thing to look at, it is true; but the cat no longer appeared to suffer any pain. As might be expected, however, it ran from me in fear whenever I came near. Why should it not run? Yet this did not fail to anger me. I felt growing inside myself a new feeling. Who has not, a hundred times, found himself doing wrong, some evil thing for no other reason than because he knows he should not? Are not we humans at all times pushed, ever driven in some unknown way to break the law just because we understand it to be the law? /n One day, in cold blood, I tied a strong rope around the cat’s neck, and taking it down into the cellar under the house I hung it from one of the wood beams above my head. I hung it there until it was dead. I hung it there with tears in my eyes, I hung it because I knew it had loved me, because I felt it had given me no reason to hurt it, because I knew that my doing so was a wrong so great, a sin so deadly that it would place my soul forever outside the reach of the love of God! /n That same night, as I lay sleeping, I heard through my open window the cries of our neighbors. I jumped from my bed and found that the entire house was filled with fire. It was only with great difficulty that my wife and I escaped. And when we were out of the house, all we could do was stand and watch it burn to the ground. I thought of the cat as I watched it burn, the cat whose dead body I had left hanging in the cellar. It seemed almost that the cat had in some mysterious way caused the house to burn so that it could make me pay for my evil act, so that it could take revenge upon me. /n Months went by, and I could not drive the thought of the cat out of my mind. One night I sat in the inn, drinking as usual. In the corner I saw a dark object that I had not seen before. I went over to see what it could be. It was a cat, a cat almost exactly like Pluto. I touched it with my hand and petted it, passing my hand softly along its back. The cat rose and pushed its back against my hand. /n Suddenly, I realized that I wanted the cat. I offered to buy it from the innkeeper, but he claimed he had never seen the animal before. As I left the inn, it followed me, and I allowed it to do so. It soon became a pet of both my wife and myself. The morning after I brought it home, however, I discovered that this cat, like Pluto, had only one eye. /n How was it possible that I had not noticed this the night before? This fact only made my wife love the cat more. But I myself found a feeling of dislike growing in me. My growing dislike of the animal only seemed to increase its love for me. It followed me, followed me everywhere, always. When I sat, it lay down under my chair. When I stood up it got between my feet and nearly made me fall. Wherever I went, it was always there. At night, I dreamed of it. And I began to hate that cat! /n One day my wife called to me from the cellar of the old building where we were now forced to live. As I went down the stairs, the cat, following me as always, ran under my feet and nearly threw me down. /n In sudden anger, I took a knife and struck wildly at the cat. Quickly my wife put out her hand and stopped my arm. This only increased my anger and, without thinking, I turned and put the knife’s point deep into her heart! She fell to the floor and died without a sound. /n I spent a few moments looking for the cat, but it was gone. And I had other things to do, for I knew I must do something with the body, and quickly. Suddenly, I noted a place in the wall of the cellar where stones had been added to the wall to cover an old fireplace which was no longer wanted. /n The walls were not very strongly built, and I found I could easily take down those stones. Behind them there was, as I knew there must be, a hole just big enough to hold the body. With much effort I put the body in and carefully put the stones back in their place. I was pleased to see that it was quite impossible for anyone to know that a single stone had been moved. /n Days passed. Still there was no cat. A few people came and asked about my wife, but I answered them easily. Then one day several officers of the police came. Certain that they could find nothing, I asked them in and went with them as they searched. /n Finally, they searched the cellar from end to end. I watched them quietly, and, as I expected, they noticed nothing. But as they started up the stairs again, I felt myself driven by some unknown inner force to let them know, to make them know, that I had won the battle. /n “The walls of this building,” I said, “are very strongly built; it is a fine old house.” And as I spoke, I struck with my stick that very place in the wall behind which was the body of my wife. Immediately I felt a cold feeling up and down my back as we heard coming out of the wall itself a horrible cry. /n For one short moment, the officers stood looking at each other. Then quickly they began to pick at the stones, and in a short time they saw before them the body of my wife, black with dried blood and smelling of decay. On the body’s head, its one eye filled with fire, its wide open mouth the color of blood, sat the cat, crying out its revenge!`
-let The_Black_DeathText = `The “Red Death” had long devastated the country. No pestilence had ever been so fatal, or so hideous. Blood was its Avator and its seal — the redness and the horror of blood. There were sharp pains, and sudden dizziness, and then profuse bleeding at the pores, with dissolution. The scarlet stains upon the body and especially upon the face of the victim, were the pest ban which shut him out from the aid and from the sympathy of his fellow-men. And the whole seizure, progress and termination of the disease, were the incidents of half an hour. /n But the Prince Prospero was happy and dauntless and sagacious. When his dominions were half depopulated, he summoned to his presence a thousand hale and light-hearted friends from among the knights and dames of his court, and with these retired to the deep seclusion of one of his castellated abbeys. This was an extensive and magnificent structure, the creation of the prince's own eccentric yet august taste. A strong and lofty wall gird led it in. This wall had gates of iron. The courtiers, having entered, brought furnaces and massy hammers and welded the bolts. They resolved to leave means neither of ingress or egress to the sudden impulses of despair or of frenzy from within. The abbey was amply provisioned. With such precautions the courtiers might bid defiance to contagion. The external world could take care of itself. In the meantime it was folly to grieve, or to think. The prince had provided all the appliances of pleasure. There were buffoons, there were improvisatori, there were ballet dancer there were musicians, there was Beauty, there was wine. All these and security within. Without was the “Red Death.” /n It was toward the close of the fifth or sixth month of his seclusion, and while the pestilence raged most furiously abroad, that the Prince Prospero entertained his thousand friends at a masked ball of the most unusual magnificence. /n It was a voluptuous scene, that masquerade. But first let me tell of the rooms in which it was held. There were seven — an imperial suite. In many palaces, however, such suites form a long and straight vista, while the folding doors slide back nearly to the walls on either hand, so that the view of the whole extent is scarcely impeded. Here the case was very different; as might have been expected from the duke's love of the bizarre. The apartments were so irregularly disposed that the vision embraced but little more than one at a time. There was a sharp turn at every twenty or thirty yards, and at each turn a novel effect. To the right and left, in the middle of each wall, a tall and narrow Gothic window looked out upon a closed corridor which pursued the windings of the suite. These windows were of stained glass whose color varied in accordance with the prevailing hue of the decorations of the chamber into which it opened. That at the eastern extremity was hung, for example, in blue — and vividly blue were its windows. The second chamber was purple in its ornaments and tapestries, and here the panes were purple. The third was green throughout, and so were the casements. The fourth was furnished and lighted with orange — the fifth with white — the sixth with violet. The seventh apartment was closely shrouded in black velvet tapestries that hung all over the ceiling and down the walls, falling in heavy folds upon a carpet of the same material and hue. But in this chamber only, the color of the windows failed to correspond with the decorations. The panes here were scarlet — a deep blood color. Now in no one of the seven apartments was there any lamp or candelabrum, amid the profusion of golden ornaments that lay scattered to and fro or depended from the roof. There was no light of any kind emanating from lamp or candle within the suite of chambers. But in the corridors that followed the suite, there stood, opposite to each window, a heavy tripod, bearing a brazier of fire, that projected its rays through the tinted glass and so glaringly illumined the room. And thus were produced a multitude of gaudy and fantastic appearances. But in the western or black chamber the effect of the fire-light that streamed upon the dark hangings through the blood-tinted panes, was ghastly in the extreme, and produced so wild a look upon the countenances of those who entered, that there were few of the company bold enough to set foot within its precincts at all. /n It was in this apartment, also, that there stood against the western wall, a gigantic clock of ebony. Its pendulum swung to and fro with a dull, heavy, monotonous clang; and when the minute-hand made the circuit of the face, and the hour was to be stricken, there came from the brazen lungs of the clock a sound which was clear and loud and deep and exceedingly musical, but of so peculiar a note and emphasis that, at each lapse of an hour, the musicians of the orchestra were constrained to pause, momentarily, in their performance, to harken to the sound; and thus the waltzers perforce ceased their evolutions; and there was a brief disconcert of the whole gay company; and, while the chimes of the clock yet rang, it was observed that the giddiest grew pale, and the more aged and sedate passed their hands over their brows as if in confused revery or meditation. But when the echoes had fully ceased, a light laughter at once pervaded the assembly; the musicians looked at each other and smiled as if at their own nervousness and folly, and made whispering vows, each to the other, that the next chiming of the clock should produce in them no similar emotion; and then, after the lapse of sixty minutes, (which embrace three thousand and six hundred seconds of the Time that flies,) there came yet another chiming of the clock, and then were the same disconcert and tremulousness and meditation as before. /n But, in spite of these things, it was a gay and magnificent revel. The tastes of the duke were peculiar. He had a fine eye for colors and effects. He disregarded the decora of mere fashion. His plans were bold and fiery, and his conceptions glowed with barbaric lustre. There are some who would have thought him mad. His followers felt that he was not. It was necessary to hear and see and touch him to be sure that he was not. /n He had directed, in great part, the moveable embellishments of the seven chambers, upon occasion of this great fete; and it was his own guiding taste which had given character to the masqueraders. Be sure they were grotesque. There were much glare and glitter and piquancy and phantasm — much of what has been since seen in “Hernani.” There were arabesque figures with unsuited limbs and appointments. There were delirious fancies such as the madman fashions. There were much of the beautiful, much of the wanton, much of the bizarre, something of the terrible, and not a little of that which might have excited disgust. To and fro in the seven chambers there stalked, in fact, a multitude of dreams. And these — the dreams — writhed in and about, taking hue from the rooms, and causing the wild music of the orchestra to seem as the echo of their steps. And, anon, there strikes the ebony clock which stands in the hall of the velvet. And then, for a moment, all is still, and all is silent save the voice of the clock. The dreams are stiff-frozen as they stand. But the echoes of the chime die away — they have endured but an instant — and a light, half-subdued laughter floats after them as they depart. And now again the music swells, and the dreams live, and writhe to and fro more merrily than ever, taking hue from the many tinted windows through which stream the rays from the tripods. But to the chamber which lies most westwardly of the seven, there are now none of the maskers who venture; for the night is waning away; and there flows a ruddier light through the blood-colored panes; and the blackness of the sable drapery appals; and to him whose foot falls upon the sable carpet, there comes from the near clock of ebony a muffled peal more solemnly emphatic than any which reaches their ears who indulge in the more remote gaieties of the other apartments. /n But these other apartments were densely crowded, and in them beat feverishly the heart of life. And the revel went whirlingly on, until at length there commenced the sounding of midnight upon the clock. And then the music ceased, as I have told; and the evolutions of the waltzers were quieted; and there was an uneasy cessation of all things as before. But now there were twelve strokes to be sounded by the bell of the clock; and thus it happened, perhaps that more of thought crept, with more of time, into the meditations of the thoughtful among those who revelled. And thus too, it happened, perhaps, that before the last echoes of the last chime had utterly sunk into silence, there were many individuals in the crowd who had found leisure to become aware of the presence of a masked figure which had arrested the attention of no single individual before. And the rumor of this new presence having spread itself whisperingly around, there arose at length from the whole company a buzz, or murmur, expressive of disapprobation and surprise — then, finally, of terror, of horror, and of disgust. /n In an assembly of phantasms such as I have painted, it may well be supposed that no ordinary appearance could have excited such sensation. In truth the masquerade license of the night was nearly unlimited; but the figure in question had out-Heroded Herod, and gone beyond the bounds of even the prince's indefinite decorum. There are chords in the hearts of the most reckless which cannot be touched without emotion. Even with the utterly lost, to whom life and death are equally jests, there are matters of which no jest can be made. The whole company, indeed, seemed now deeply to feel that in the costume and bearing of the stranger neither wit nor propriety existed. The figure was tall and gaunt, and shrouded from head to foot in the habiliments of the grave. The mask which concealed the visage was made so nearly to resemble the countenance of a stiffened corpse that the closest scrutiny must have had difficulty in detecting the cheat. And yet all this might have been endured, if not approved, by the mad revellers around. But the mummer had gone so far as to assume the type of the Red Death. His vesture was dabbled in blood — and his broad brow, with all the features of the face, was besprinkled with the scarlet horror. /n When the eyes of Prince Prospero fell upon this spectral image (which with a slow and solemn movement, as if more fully to sustain its role, stalked to and fro among the waltzers) he was seen to be convulsed, in the first moment with a strong shudder either of terror or distaste; but, in the next, his brow reddened with rage. /n “Who dares?” he demanded hoarsely of the courtiers who stood near him — “who dares insult us with this blasphemous mockery? Seize him and unmask him — that we may know whom we have to hang at sunrise, from the battlements!” /n It was in the eastern or blue chamber in which stood the Prince Prospero as he uttered these words. They rang throughout the seven rooms loudly and clearly — for the prince was a bold and robust man, and the music had become hushed at the waving of his hand. /n It was in the blue room where stood the prince, with a group of pale courtiers by his side. At first, as he spoke, there was a slight rushing movement of this group in the direction of the intruder, who, at the moment was also near at hand, and now, with deliberate and stately step, made closer approach to the speaker. But from a certain nameless awe with which the mad assumptions of the mummer had inspired the whole party, there were found none who put forth hand to seize him; so that, unimpeded, he passed within a yard of the prince's person; and, while the vast assembly, as if with one impulse, shrank from the centres of the rooms to the walls, he made his way uninterruptedly, but with the same solemn and measured step which had distinguished him from the first, through the blue chamber to the purple — through the purple to the green — through the green to the orange — through this again to the white — and even thence to the violet, ere a decided movement had been made to arrest him. It was then, however, that the Prince Prospero, maddening with rage and the shame of his own momentary cowardice, rushed hurriedly through the six chambers, while none followed him on account of a deadly terror that had seized upon all. He bore aloft a drawn dagger, and had approached, in rapid impetuosity, to within three or four feet of the retreating figure, when the latter, having attained the extremity of the velvet apartment, turned suddenly and confronted his pursuer. There was a sharp cry — and the dagger dropped gleaming upon the sable carpet, upon which, instantly afterwards, fell prostrate in death the Prince Prospero. Then, summoning the wild courage of despair, a throng of the revellers at once threw themselves into the black apartment, and, seizing the mummer, whose tall figure stood erect and motionless within the shadow of the ebony clock, gasped in unutterable horror at finding the grave cerements and corpse-like mask which they handled with so violent a rudeness, untenanted by any tangible form. /n And now was acknowledged the presence of the Red Death. He had come like a thief in the night. And one by one dropped the revellers in the blood-bedewed halls of their revel, and died each in the despairing posture of his fall. And the life of the ebony clock went out with that of the last of the gay. And the flames of the tripods expired. And Darkness and Decay and the Red Death held illimitable dominion over all.`
-
-
-let PeterRabbit = {
-    Text: PeterRabbitText,
-    SearchTerms: ['children', 'rabbit', 'peter', 'peter rabbit', 'childrens', 'potter', 'beatrix potter', 'bunny', 'early', 'easy', 'early readers'],
-    Cover: peterRabbitCover,
-    Title: 'PeterRabbit'
-}
-
-let Tortoise_and_the_Hare = {
-    Text: Tortoise_and_the_HareText,
-    SearchTerms: ['tortoise', 'turtle', 'tortoise and the hare', 'hare', 'race', 'children', 'childrens', 'rabbit', 'bunny', 'early', 'easy', 'early readers'],
-    Cover: tortHareCover,
-    Title: 'Tortoise_and_the_Hare'
-}
-
-let Little_Red_Hen = {
-    Text: Little_Red_HenText,
-    SearchTerms: ['hen', 'red', 'chicken', 'the little red hen', 'little red hen', 'children', 'childrens', 'early', 'easy', 'early readers'],
-    Cover: littleRedHenCover,
-    Title: 'Little_Red_Hen'
-}
-
-let The_Black_Cat = {
-    Text: The_Black_CatText,
-    SearchTerms: ['cat', 'black cat', 'edgar allen poe', 'the black cat', 'poe', 'advanced readers'],
-    Cover: theBlackCatCover,
-    Title: 'The_Black_Cat'
-}
-
-let The_Masque_Of_The_Red_Death = {
-    Text: The_Black_DeathText,
-    SearchTerms: ['death', 'black death', 'edgar allen poe', 'the masque of the red death', 'poe', 'the mask of the red death', 'advanced readers'],
-    Cover: RedDeathCover,
-    Title: 'The_Masque_Of_The_Red_Death'
-}
-
-export default {
-    PeterRabbit: PeterRabbit,
-    Tortoise_and_the_Hare: Tortoise_and_the_Hare,
-    Little_Red_Hen: Little_Red_Hen,
-    The_Black_Cat: The_Black_Cat,
-    The_Masque_Of_The_Red_Death: The_Masque_Of_The_Red_Death
-}
