+++ src/assets/The_Black_CatText.txt
@@ -1,0 +1,1 @@
+Tomorrow I die. Tomorrow I die, and today I want to tell the world what happened and thus perhaps free my soul from the horrible weight which lies upon it. /n But listen! Listen, and you shall hear how I have been destroyed. /n When I was a child, I had a natural goodness of soul which led me to love animals — all kinds of animals, but especially those animals we call pets, animals which have learned to live with men and share their homes with them. There is something in the love of these animals which speaks directly to the heart of the man who has learned from experience how uncertain and changeable is the love of other men. /n I was quite young when I married. You will understand the joy I felt to find that my wife shared with me my love for animals. Quickly she got for us several pets of the most likeable kind. We had birds, some goldfish, a fine dog, and a cat. /n The cat was a beautiful animal, of unusually large size, and entirely black. I named the cat Pluto, and it was the pet I liked best. I alone fed it, and it followed me all around the house. It was even with difficulty that I stopped it from following me through the streets. /n Our friendship lasted, in this manner, for several years, during which, however, my own character became greatly changed. I began to drink too much wine and other strong drinks. /n As the days passed I became less loving in my manner; I became quick to anger; I forgot how to smile and laugh. My wife — yes, and my pets, too, all except the cat — were made to feel the change in my character. /n One night I came home quite late from the inn, where I now spent more and more time drinking. Walking with uncertain step, I made my way with effort into the house. As I entered I saw — or thought I saw — that Pluto, the cat, was trying to stay out of my way, to avoid me. This action, by an animal which I had thought still loved me, made me angry beyond reason. My soul seemed to fly from my body. I took a small knife out of my coat and opened it. Then I took the poor animal by the neck and with one quick movement I cut out one of its fear-filled eyes! /n Slowly the cat got well. The hole where its eye had been was not a pretty thing to look at, it is true; but the cat no longer appeared to suffer any pain. As might be expected, however, it ran from me in fear whenever I came near. Why should it not run? Yet this did not fail to anger me. I felt growing inside myself a new feeling. Who has not, a hundred times, found himself doing wrong, some evil thing for no other reason than because he knows he should not? Are not we humans at all times pushed, ever driven in some unknown way to break the law just because we understand it to be the law? /n One day, in cold blood, I tied a strong rope around the cat’s neck, and taking it down into the cellar under the house I hung it from one of the wood beams above my head. I hung it there until it was dead. I hung it there with tears in my eyes, I hung it because I knew it had loved me, because I felt it had given me no reason to hurt it, because I knew that my doing so was a wrong so great, a sin so deadly that it would place my soul forever outside the reach of the love of God! /n That same night, as I lay sleeping, I heard through my open window the cries of our neighbors. I jumped from my bed and found that the entire house was filled with fire. It was only with great difficulty that my wife and I escaped. And when we were out of the house, all we could do was stand and watch it burn to the ground. I thought of the cat as I watched it burn, the cat whose dead body I had left hanging in the cellar. It seemed almost that the cat had in some mysterious way caused the house to burn so that it could make me pay for my evil act, so that it could take revenge upon me. /n Months went by, and I could not drive the thought of the cat out of my mind. One night I sat in the inn, drinking as usual. In the corner I saw a dark object that I had not seen before. I went over to see what it could be. It was a cat, a cat almost exactly like Pluto. I touched it with my hand and petted it, passing my hand softly along its back. The cat rose and pushed its back against my hand. /n Suddenly, I realized that I wanted the cat. I offered to buy it from the innkeeper, but he claimed he had never seen the animal before. As I left the inn, it followed me, and I allowed it to do so. It soon became a pet of both my wife and myself. The morning after I brought it home, however, I discovered that this cat, like Pluto, had only one eye. /n How was it possible that I had not noticed this the night before? This fact only made my wife love the cat more. But I myself found a feeling of dislike growing in me. My growing dislike of the animal only seemed to increase its love for me. It followed me, followed me everywhere, always. When I sat, it lay down under my chair. When I stood up it got between my feet and nearly made me fall. Wherever I went, it was always there. At night, I dreamed of it. And I began to hate that cat! /n One day my wife called to me from the cellar of the old building where we were now forced to live. As I went down the stairs, the cat, following me as always, ran under my feet and nearly threw me down. /n In sudden anger, I took a knife and struck wildly at the cat. Quickly my wife put out her hand and stopped my arm. This only increased my anger and, without thinking, I turned and put the knife’s point deep into her heart! She fell to the floor and died without a sound. /n I spent a few moments looking for the cat, but it was gone. And I had other things to do, for I knew I must do something with the body, and quickly. Suddenly, I noted a place in the wall of the cellar where stones had been added to the wall to cover an old fireplace which was no longer wanted. /n The walls were not very strongly built, and I found I could easily take down those stones. Behind them there was, as I knew there must be, a hole just big enough to hold the body. With much effort I put the body in and carefully put the stones back in their place. I was pleased to see that it was quite impossible for anyone to know that a single stone had been moved. /n Days passed. Still there was no cat. A few people came and asked about my wife, but I answered them easily. Then one day several officers of the police came. Certain that they could find nothing, I asked them in and went with them as they searched. /n Finally, they searched the cellar from end to end. I watched them quietly, and, as I expected, they noticed nothing. But as they started up the stairs again, I felt myself driven by some unknown inner force to let them know, to make them know, that I had won the battle. /n “The walls of this building,” I said, “are very strongly built; it is a fine old house.” And as I spoke, I struck with my stick that very place in the wall behind which was the body of my wife. Immediately I felt a cold feeling up and down my back as we heard coming out of the wall itself a horrible cry. /n For one short moment, the officers stood looking at each other. Then quickly they began to pick at the stones, and in a short time they saw before them the body of my wife, black with dried blood and smelling of decay. On the body’s head, its one eye filled with fire, its wide open mouth the color of blood, sat the cat, crying out its revenge!
