+++ src/assets/Tortoise_and_the_HareText.txt
@@ -1,0 +1,1 @@
+One day a hare was bragging about how fast he could run. He bragged and bragged and even laughed at the tortoise, who was so slow. /n The tortoise stretched out his long neck and challenged the hare to a race, which, of course, made the hare laugh. /n "My, my, what a joke!" thought the hare. "A race, indeed, a race. Oh! what fun! My, my! a race, of course, Mr. Tortoise, we shall race!" said the hare. /n The forest animals met and mapped out the course. The race began, and the hare, being such a swift runner, soon left the tortoise far behind. /n About halfway through the course, it occurred to the hare that he had plenty of time to beat the slow trodden tortoise. "Oh, my!" thought the hare, "I have plenty of time to play in the meadow here." And so he did. /n After the hare finished playing, he decided that he had time to take a little nap. "I have plenty of time to beat that tortoise," he thought. And he cuddle up against a tree and dozed. /n The tortoise, in the meantime, continued to plod on, albeit, it ever so slowly. He never stopped, but took one good step after another. /n The hare finally woke from his nap. "Time to get going," he thought. And off he went faster than he had ever run before! He dashed as quickly as anyone ever could up to the finish line, where he met the tortoise, who was patiently awaiting his arrival. /n Slow and steady wins the race.
